requirements.txt
deepagents
langchain-openai
langgraph
tavily-python
croniter
langchain-mcp-adapters
lark-oapi>=1.4.8
google-genai
Pillow
DrissionPage
sqlite-vec
openai
requests
pandas
openpyxl
pypdf
python-docx
